+++ CHANGELOG.md
@@ -23,6 +23,8 @@ All notable changes to this project will be documented in this file.
 - Added example `draw_shapes_shader.rs` to show how to set a custom pipeline drawing shapes.
 - Renamed `draw_shader.rs` to `draw_image_shader.rs`
 - Added `Graphics::stats() -> GpuStats` to get more info about what the GPU did the last frame. 
+- Added new texture formats. `TextureFormat::R16Uint`, `R32Uint`, `R32Float`. 
+- New example `renderer_texture_r32` to show how to use new texture types.
 
 ## v0.7.1 - 08/10/2022